APPS-3470 switch stroke dashing from float to double

+++ retex/renderer-web/src/main/java/com/himamis/retex/renderer/web/graphics/GraphicsFactoryGWT.java
@@ -87,7 +87,7 @@ public class GraphicsFactoryGWT extends GraphicsFactory {
 	}
 
 	@Override
-	public Stroke createBasicStroke(double width, float[] dashes) {
+	public Stroke createBasicStroke(double width, double[] dashes) {
 		return new BasicStrokeW(width, dashes);
 	}
